cat_recognizer/requirements.txt
numpy
tensorflow-cpu
scikit-learn
Pillow